nbip_server/requirements.txt
Django ==1.6.1
django-bootstrap-form ==3.1
django-registration ==1.0
